feat: Refactor chat components to use new ComposerMenu and ComposerPreferences hooks
- Implemented useComposerMenu for managing menu states and positioning in ModelReasoningPicker and PermissionModePicker.
- Removed ReasoningLevelPicker component and integrated its functionality into existing components.
- Added ComposerAttachmentMenu for file and directory attachment options.
- Introduced ComposerContextUsageIndicator to display context usage information.
- Created ComposerModeControls to encapsulate mode selection controls.
- Added DefaultConnectionConfirmDialog for confirming default connection settings.
- Implemented useComposerPreferences for managing agent mode and reasoning level preferences in local storage.
