+++ src/i18n/app-messages.en.ts
@@ -318,10 +318,8 @@ export const enMessages = {
   "chat.contextUsagePanelInputLimit": "Model input limit: {limit}",
   "chat.contextUsagePanelWindow": "Model window: {limit}",
   "chat.contextUsagePanelTokensUnknown": "{used} tokens used",
-  "chat.modelPicker": "Switch model",
   "chat.modelReasoningPicker": "Switch model and reasoning",
   "chat.modelSection": "Model",
-  "chat.modelBuiltIn": "Built-in models",
   "chat.modelCustom": "Custom models",
   "chat.modelAdd": "Configure custom model",
   "chat.modelDelete": "Delete custom model",
@@ -347,7 +345,6 @@ export const enMessages = {
   "chat.agentModePicker": "Switch mode",
   "chat.agentModeBuild": "Execute",
   "chat.agentModePlan": "Plan",
-  "chat.reasoningLevelPicker": "Reasoning level",
   "chat.reasoningSection": "Reasoning",
   "chat.reasoningLevelDefault": "Default",
   "chat.reasoningLevelLow": "Low",
