+++ src/i18n/app-messages.zh.ts
@@ -305,10 +305,8 @@ export const zhCNMessages = {
   "chat.contextUsagePanelInputLimit": "模型输入上限：{limit}",
   "chat.contextUsagePanelWindow": "模型窗口：{limit}",
   "chat.contextUsagePanelTokensUnknown": "已用 {used} 标记",
-  "chat.modelPicker": "切换模型",
   "chat.modelReasoningPicker": "切换模型和推理",
   "chat.modelSection": "模型",
-  "chat.modelBuiltIn": "内置模型",
   "chat.modelCustom": "自定义模型",
   "chat.modelAdd": "配置自定义模型",
   "chat.modelDelete": "删除自定义模型",
@@ -334,7 +332,6 @@ export const zhCNMessages = {
   "chat.agentModePicker": "切换模式",
   "chat.agentModeBuild": "执行",
   "chat.agentModePlan": "计划",
-  "chat.reasoningLevelPicker": "推理等级",
   "chat.reasoningSection": "推理",
   "chat.reasoningLevelDefault": "默认",
   "chat.reasoningLevelLow": "低",
