+++ src/routes/Chat/ModelControls.tsx
@@ -1,5 +1,3 @@
 export { AddCustomModelDialog } from "./AddCustomModelDialog.tsx"
 export { AgentModePicker } from "./AgentModePicker.tsx"
-export { ModelPicker } from "./ModelPicker.tsx"
 export { ModelReasoningPicker } from "./ModelReasoningPicker.tsx"
-export { ReasoningLevelPicker } from "./ReasoningLevelPicker.tsx"
